🚚(config) refactor configuration files and templates
We renamed the "config" directory to "configmap" to better describe
what it is in charge of and remove confusion with edxapp's "config"
folder. The settings.yml files specific to customers are replaced
by a template in which variables specific to a customer will be
called. This will be more powerful, simpler and make sure all
elements of configuration are also available as variables in a
DRY way.