🐳(ansible) allow running our Ansible playbooks in a container
Arnold's playbooks will be played by developpers on their laptop
but also by GitLab for the CI/CD and by OpenShift for init
containers. The best way to make it seemless is to do it with
Docker. We include in the image a customization of Ansible to
allow setting the vault pass as an environment variable.