👷(gitlab) add base CI for Arnold
This is a draft CI proposal for arnold that:

- builds Arnold's docker image
- tests built image with the oc command
- runs the deploy.yml ansible playbook

Disclaimer: this last job (deploy) fails

At the time of writing, this job is a POC to ensure that we can run
ansible playbooks within GitLabCI using our OpenShift Service Account
(deployment bot). Even if the playbook run fails, it runs; this is what
we wanted to achieve. The next step is to improve our deployment
strategy (playbook).

TODO:

- activate overlay2 docker driver
- Unset INSECURE_SKIP_TLS_VERIFY when switching to the production server
- make the deploy job generic