🚧(edxapp) temporary inactivate edxapp objects
At the time of writing our goal is to be able to deploy richie for a
particular customer/environment. To prevent side effects from other
services related to edxapp, we temporary inactivate them.

Nota bene: we needed to rename cms/lms nginx configuration files to
prevent related configmap creation. Nginx refuses to start when a
target reverse proxy is not responding.