🔧(environment) switch default to developement
We now use the 'development' env_type as the default environment type
and 'Development' as the default Django configuration. This makes sense
as when willing to deploy on a particular environment, this should be
explicit. Moreover, I'm too lazy during the development phase to
override the env_type environment variable everytime I type ansible
commands.