🔥(projects) limit default project to just hello
The default project should be as simple as possible.
The openedx and richie installations are saved and
will be added in a future commit in a separate
"examples" folder containing a "templates" and a
"group_var" directories that can be mounted on the
root ones when running a docker container in Arnold.