🎨(app) make nginx an app-specific service
Until now, we considered that nginx may be a single service used by many
apps at the same time. Even if this pattern is quite reasonable, it
offers less configurability if we want to deploy not all but a selection
of applications Arnold can play with. Hence we decided to split the
monolithic nginx service into multiple nginx services (one per
application). It means that we had to:

* split the common/nginx svc
* split the common/nginx dc
* split nginx configMaps
* redefine app-specific routes to point to the appropriate nginx svc
* improve route aliases definition to be app specific