✨(job) ease job restart
Add a unique job_stamp to jobs so that they can be restarted on failure
instead of cleaning them in the console and then re-created.

I choose to keep using the create_object.yml playbook to restart jobs
instead of creating a new one, as almost the entire playbook will be
duplicated.

A bin/job sugar script has also been added to ease running jobs in
development.