✨(playbook) add apps metadata from group_vars
We now allow to define an "apps" var for a customer or an environment to
add metadata to applications and their services.

Known limitation: for now, we only support the merge strategy between
discovered apps and groups_vars defined metadata. This can be refined
later if there is a need to strictly override some definitions (and not
only add metadata).