🐛(deploy) fix old stack removal
Previous old stack deletion implementation had a greedy behavior and was
also deleting newly deployed services (sic!). Current implementation is
first based on querying k8s API based on objects labels and not their
full definition. Knowing that every object we create during the
deployment is labelled with the appropriate deployment_stamp, this is
the safest way to proceed.