🐛(playbook) retry to create project on failure
When dealing with large projects (with many objects), the project
deletion may be longer than for small projects. In some cases, when
trying to create a new project with the same name as a recently
deleted one, OpenShift's API refuses to create this new project as it is
conflicting with the older one (in fact it is still working on the old
project clean up).

We've implemented a wait loop that ensures that the project has been
created after multiple retries.