🔧(vars) update default vars for Richie
Each application that can be deployed with Arnold should bundle default
configuration variables that can be overridden for a customer, an
env_type or a combination of both.

Convention: each app-specific variable should be prefixed by the name of
the application, e.g. 'richie_'.

In this work, we've moved default variables for Richie in the
application itself: "apps/richie/vars/all/main.yml" and renamed them to
respect the aforementioned convention.