✨(jobs) allow ordered sequential jobs execution
We are now able to run jobs in a particular order depending on their
template file name: jobs are sorted by their filename for an application
(all services), e.g.:

1. apps/edxapp/templates/cms/job_01_collectstatic.yml.j2
2. apps/edxapp/templates/lms/job_02_collectstatic.yml.j2
3. ...

And jobs are now executed sequentially, e.g. a new job cannot be
submitted if the previous one in the list did not complete.