✨(edxapp) add redis app
We are going to use redis for Celery instead of RabbitMQ.
Let's add a Redis running the latest stable stretch image
of Redis. It works in OpenShift without any modification.
We declare it as a separate app so that we can exclude it
from our frequent updates of the other applications. Note
that blue/green is not activated on Redis because only
one container at a time should be mounted on the Redis
data volume.