🐛(apps) load customer/env_types vars before apps management
To be able to define the apps variable for a customer/env_type, vars
should be loaded before trying to create/merge the apps variable.

KNOWN ISSUE: as we load default vars for each app later, they are no
longer overriden. We should set default vars from the app definition
only if they are not already defined.