✨(edxapp) add theme to edxapp LMS
The theme is added in a build step specific to each customer upon
deployment. The same build will be used to add functionalities
on the common image that are either specific to our OpenShift
deployment (e.g. adding Datadog monitoring) our specific to
a customer (theme).

It is a good idea to use a common image stream for the CMS and the
LMS even if no customization is made because it allows keeping a
copy of the image in our local registry and avoid downloading it
from Doker Hub upon each deployment.