🐛(richie) add node affinity rule to run ES pods
Running ElasticSearch in a pod requires that the node running it has a custom
kernel parameter (vm/max_map_count=262144). As it's a potential security issue
and can lead to unexpected side effects, required changes have been made on a
few nodes labelled with "node_with_max_map_count=true". Thus, we need to target
those nodes for this deployment.