⚡️(templates) filter app rendered empty templates
Our wait loop during deployment is based on the fact that for every
DeploymentConfig, we should have a running pod. If some DCs are not
created under certain conditions (e.g. env_type), our wait loop is
infinite and thus broken. To solve this issue, we filter templates that
are empty when rendered so that the object list we cook only contains
objects that will get created.