✨(email) add a mailcatcher app to handle emails in each project
On environments other than production, we don't want our
applications to send real emails but we want to be able to
read and act on the emails sent. The best practice to do this
is to install a mail catcher. We will run one instance of
mailcatcher in each project (shared by all apps in this project).