🐛(edxapp) stop overridding SITE_NAME
The version of edxapp hawthorn.1-1.0.4 sets SITE_NAME to
the LMS hosts (resp. CMS) for the lms (resp cms) app. so
we don't need to set it here. It was not set to the correct
value as edxapp expects a domain name, not a display name,
and uses it to build links to the app in the emails it
sends.