⚡️(richie) add new application secrets
Secrets are now stored separately in a dedicated template for each
service. Vars are defined in the vault, decrypted on the fly and
injected in the template of the secret. Note that secret name and
versionning should be done using new variables defined for the
application, e.g. {{ app.name }}_{{ app_service }}_secret_name.