🚸(bin) add script to build images as it is now separate from deploy
We recently took the image build out of deployment because we were
making many more deployments than image builds and the image build
is consuming a lot of resources. We forgot to add a helper for this
command forcing users to either use bin/bootstrap or type the whole
docker run command.