⬆️(edxapp) update edxapp default image to hawthorn.1-2.0.0
This image is a complete refactorization of the build process and
fixes issues on settings that were blocking some deployments like
mongodb port numbers not being formatted to integers. The
webpack-stats.json is now already copied to a location collected
by collectstatic so we should not try to copy it anymore in Arnold.