🐛(plugin) fix blue_green_hosts filter output
Instead of returning a list that will be harder to parse once converted
to a string, make the blue_green_hosts filter return a string with
comma-separated items. By doing so, if an environment variable is set
with the returned value, most application will be able to extract host
names.