🔥(core) remove useless apt requirements
Storing 4 pinned dependencies in a separated file seems overkill for an
image that will run locally in the developers' machine.

Moreover, pinned dependencies are quickly deprecated when the base
official image's repositories are updated. This breaks the build for no
relevant reason, so let's stick to a more classical approach.