🐛(nginx) fix nginx config to work behind the TLS termination proxy
Nginx acts as a proxy and was overriding the value of the X-Forwarded-
headers. The problem is that our applications are running in OpenShift
behind an HAProxy that handles the TLS termination. This proxy has
already set the X-Forwarded- headers with the values received from the
client, so we should respect them and not override them with the values
in the request hitting nginx because they come from the HAProxy,
not the client.