✨(security) automate http basic auth htpasswd file generation
In some cases, we must be able to protect a publicly deployed service
behind an http basic auth, e.g. to prevent search engines bots crawling
or restrict access to a confidential application instance.

We have automated the creation of a htpasswd file for a list of users.
One file is created per customer/environment/application triplet. Next
step is to create a secret from this file; this secret will be
configured and injected in application's nginx pods.