✨(edxapp) add new route for preview url
The preview url is used to allow instructors to review a course that is
now published on the LMS. It is an accessible by intructors of the course
and unlock displaying of the draft items in the course. Since this is
using a specific domain, we must create a specific route and add it to
allowed hosts.