✨(jobs) add support for pre/post deployment jobs
Some jobs require to be run before deploying an application, hence we
introduced a way to filter jobs and execute them before or after
creating deployments.

Filtering is based on the "job_type" label, pre-deployment jobs must be
labelled with job_type="pre" and post-deployment jobs must be labelled
with job_type="post".