✨(volumes) add create_volumes playbook
When developing a new feature that requires a new volumes, one may need
to create this new volume on a live project that already have been
initialized.

We've added a create_volumes playbook that can be run during project's
initialization and on a live project. We've removed the "force: true"
flag on volumes creation which does not seem to be required.