✨(volumes) make PV reclaim policy configurable
By default, Kubernetes deletes a PV when a PVC is deleted by a user,
leading to irreversible data loss. In some cases (_e.g._ environments),
we want to keep a volume even if the corresponding PVC has been deleted.

The default reclaim policy for production is "Retain" and "Delete" for
all other environments.