🐛(redirect) fix routes base template path
Since we now declare explicitly routes for an application by creating
one (or many) route*.yml.j2 template(s) for each application service,
the redirect app base template was consider as a route and not a
generic route template used for redirections. Now that this template is
considered as private (prefixed with an underscore), it is ignored when
trying to create or patch routes.