⚡️(edxapp) add health check probes
We need a way to automate container restart when it is failing. This
could be achieved thanks to the health check probes that kubernetes
offers.

For edxapp, we've used openedx's native healthcheck view for both
readiness and liveness probes of wsgi services (CMS + LMS). For celery
workers, we take advantage of the ping management command to check
container status and readiness.