⚡️(redis) add health check probes
It is commonly assumed that having a response to the "redis-cli info"
command means that the server is up and running, hence, we use this
command for our readiness and liveness probes.

We've decided to wait 60 seconds before starting to check application's
liveness, supposing that if it's ready the application has already
booted.