⚡️(playbook) restrict apps_filter to switch/deploy
The switch playbook is supposed to run against a single application at a
time due to unexpected results on non-blue-green apps or not newly
deployed apps. Thus, as for the deploy playbook, we now require that the
apps_filter is defined when invoking the switch playbook.

For deployment and switch, we only require the apps_filter to be defined
if the current project has multiple activated apps.