♻️(core) refactor app deletion
Since Ansible 2.7 upgrade, app deletion (during blue/green switch or
next stack deployment) was partially working. We've decided to refactor
this feature thanks to the k8s_facts module and a less convoluted
approach. We've added a new step in the CI to test it, and also
refactored CI scripts to make them more generic.