🐛(django) flag collectstatic jobs as "pre" jobs
When using the ManifestStaticFilesStorage, the collectstatic management
command will post-process static files to add a md5 hash of the file in
its name. If this hash changes during the collectstatic job, the django
application pod will serve the wrong version of this file.

Thus we need to perform the collecstatic job **before** deploying the
new version of the application.