✨(marsha) allow to configure the command running marsha contrainer
Gunicorn in docker need few modifications to run well an application
(see https://pythonspeed.com/articles/gunicorn-in-docker/). For marsha
we want to apply some of them rule, the easiest and flexible way to do
this is to override the command running in the container and set it in a
variable.