♻️(richie) allow deactivating volumes for the static and media files
For most of our projects, we use AWS S3 & CloudFront to store and
serve our static and media files. In this case, creating volumes in
OpenShift is useless. After this PR, creating volumes remains the
default behavior as it is simpler and works out of the box. Also,
relying on AWS requires a Terraform project that is not in richie's
sandbox yet.