✨(redis) create a config map to manage redis configuration
We want to be able to customize redis configuration. It is possible to
start redis server by specifying it a configuration file. This
configuration file is loaded using a config map and a volume in the DC.
We also have to change the command to start the container to specify the
configuration file.