✨(forum) configure mongo to use replicaset and read_preference settings
We want to connect to a mongo cluster using replicaset and
read_preference features. In the forum we use a mongo URI, we created a
macro responsible to compute this URI depending if we are or not in a
trashable environment. Replicaset and read_preference are optional and
read_preference will not be used if replicaset setting is not set.