♻️(registry) use local lookupPolicy
ImageStream can use first the local registry and then fetch on
docker hub if the image is not present. To do that the local
lookupPolicy must be set to true and there is no need anymore to refer
to the internal docker registry in the image name used in every
DeploymentConfig or jobs.