✨(playbook) introduce static services
Add support  for static services, with  fixed name, in addition  to the services
deployed with  a deployment stamp in  their name. It allows  applications inside
the cluster to use the services exposed by other applications, with a consistent
fqdn.

BREAKING  CHANGE :  blue/green related  tasks are  now updating  static services
instead of routes