✨(marsha) filter access to /admin with ips in a whitelist
We want to give access to /admin path to a restricted list of ips we
want to control. By default the variable marsha_nginx_admin_ip_withelist
is set with an empty list and if ips are added to this list they will be
used to restrict the access of /admin to this given list.