✨(learninglocker) filter access to UI with ips in a whitelist
We want to give access to learninglocker UI to a restricted list of ips we
want to control. By default the variable learninglocker_nginx_admin_ip_withelist
is set with an empty list and if ips are added to this list they will be
used to restrict the access of the UI to this given list.