✨(apps) add elasticsearch application
Add  an  independant  elasticsearch   app  with  a  configurable  high
availability cluster. It is designed  to run without persistent volume
and it  supports rolling updates without  any data loss. On  the other
hand,  a full  restart  of  the ES  cluster  (without rolling  update)
results in data loss and requires to recreate indexes.