♻️(task) delete CronJob resource in delete_app task
CronJob resource shloud be deleted like any other resources when the
delete_app task is executed. We faced an issue doing this, `CronJob`
doesn't use the apiVersion `v1` but `v1beta1` and this value was
hardcoded in our code. The apiVersion value is now set in the delete_app
task and used by delete_app_objects_kind task.