✅(scripts) add argument to ci-test-route to set the max retries
The ci-test-route  script has  a hardcoded value  (5) for  the maximum
number of retries to get the  expected response from the server before
failing.  This  number  is   too  low  for  the  test-bootstrap-edxapp
application and  generates a lot  of false negative  in the CI.  A new
argument  to the  script is  added to  specify the  maximum number  of
retries.