🐛(edxapp) reinstate static volume on edxapp as temporary fix
Arnold v5.0.0 introduced an nginx docker image that includes static
files for performance improvement. This feature was broken because
the edxapp container could not access static files anymore and it
is required by the CachedStaticFilesStorage backend to calculate file
hashes. Bringing back the static volume will allow running
collectstatic manually in edxapp containers so that Django can access
the files at run time. This buys us time to work on a real fix.