✨(learninglocker) configure and use pm2 to run learninglocker
Recent docker images of learninglocker have pm2 installed. We want to
use this process manager in Arnold. For each service we created a
configuration file containing information on how to run each
application.
Also pm2 writes files and direcotries in the HOME user but by default on
OpenShift the HOME environment variable is set to `/` which is obviously
not writable for an un-privileged user. We need to set this variable
with a directory the user can writes in.