♻️(edxapp) refactor for pre-built edxapp images with static symlinks
We decided to pre-build edxapp images outside of Arnold. The pre-built
images should be ready to run ie with applied custom theme if any and
including static files as symlinks to avoid having to collectstatic
files in a volume (cf github.com/openfun/openedx-docker). As a result,
we can completely remove everything related to building custom images
and collecting static files from Arnold.