🔧(edxapp) limit static files cache to 1 day
When referring to static files in a requirejs statement, Open edX does
not use the hashed file name but the original file name of the static
file. These files may change when we upgrade the base Open edX image of
one of our site, so we should keep cache expiry low.