✨(apps) choose to enable acme based on route prefix
With acme v2, each time it tries to resolve a challenge a new route is
created and an associated pod is also created. In some cases, routes
next and previous exist but are not used, DNS is not created for example,
the challenge is never resolved and routes and pods are still living in
the project consuming resources for nothing. We introduced a new
variable acme_enabled_route_prefix allowing to choose on which route
acme will generate a certificate based on their prefix.