🐛(playbook) fix the `delete_app` task to avoid timeout
When executing a `clean` on a project having `Cronjob` objects, it
fails with a timeout on the step "Wait for objects of kind Pod to be
deleted". The reason is that we are trying to delete pods before
cronjobs. As soon as we delete a pod related to a cronjob (or a job),
it is automatically re-created. And we can wait forever for the pods
to be deleted. The pods must be deleted last.