👷(project) enhance CI build by building only updated apps
This commit introduce a new way of running the CI.
Before the CI will run, it will check that the config.yml is
updated regarding the current changes in the branch.
It the config.yml is not up to date, the build will stop
very early. It he file us up to date, then only apps that have
changes should be run.