⚡️(project) do not set apps in env_types vars
Enforcing the `apps` variable definition in an environment type means
that it does not default to available apps. This is annoying while
developing a tray because one need to override the `apps` variable
definition in the customer/env_type variables to make our tray
deployable. More than that it seems more maintenable as is.