✨(cli) improve container volumes management
env_types:

When we need to override default env_type vars definition, it's usefull
to redefine them in the local arnold project. The default behavior is
now to mount the projetc's `groups_vars/env_type` directory if it has
been defined locally.

volumes:

More generally, when working with Arnold, we might want to add specific
resources or override a container directory (e.g. `apps/`) in the
execution context. To fill the gap, we introduced a new `--extra-volume`
option that allows this feature.

k8s_domain:

Arnold CLI now allows to set the K8S_DOMAIN context by using an option
instead of an environment variable.