🐛(cli) add missing export to K8S_DOMAIN in arnold CLI
When the `--k8s-domain` is not provided to arnold CLI, a default value
is computed with the local IP address. This default value is not
exported to the docker container running ansible tasks though the
K8S_DOMAIN environment variable. As a result the base domain used in
routes is invalid.