🏗(cli) use k8s instead of OKD in arnold CLI
We want to replace OKD with Kubernetes. The authentication mechanism
is sligthly different, because the commands `oc login` and `oc whoami
-t` do not exist with `kubectl`.

Arnold CLI has been updated to discover authentication information in
k8s configuration and service account secrets, as explained here :
https://kubernetes.io/docs/tasks/administer-cluster/access-cluster-api/

Note that it is now mandatory to authenticate through a service
account to use arnold. Authenticating with a regular user is not
possible anymore, because it can be done in multiple ways (e.g. with
OpenID, via a kubectl plugin that need user interaction) and this
cannot be executed inside the docker container.

For local development purpose, the `oc cluster` command has been
replaced with k3d, an utility to start a k3s cluster within docker. By
default, k3d installs traefik as the default ingress controller. We
chose to replace it by the ingress-nginx controller instead, because
it's a popular choice of ingress controller and it is officially
supported by the kubernetes project. And unlike traeffik, by default,
the nginx controller provides a default self-signed certificate if
cert-manager is not installed.