🔨(k3d) add support for RWX volumes in k3d
Some Arnold apps require PersistentVolumeClaim with ReadWriteMany
(RWX) acces mode. But out of the box, k3d is bundled with a
"local-path" StorageClass and a dynamic provisionner that can only
provision PersistentVolumes with ReadWriteOnce access mode. As a
result, our volume claims cannot be satisfied.

To solve that, since it concerns only the dev and CI environments, the
`bin/init-cluster` script pre-provision a bunch of PersistentVolume
each time we start the k3d cluster. These pre-provisionned PVs are
only virtual and do not consume or reserve any resource until they are
bound to a PVC.

The data stored on these PVs are written in a docker volume.