✨(core) allow to specify uid of running containers
By default, kubernetes will run containers with the USER specified in
the docker image. And this USER can be root. We want to run containers
as non-root user, like Openshift was doing by default.

Kubernetes allow to specify a SecurityContext for resources that will
create containers (e.g. Pods, Deployments, Jobs...). SecurityContext
has a RunAsUser option which allows to run containers with the
specified UID.

A new ansible variable `container_uid` has been introduced to specify
the UID that will be used to run containers. We recommend to specify a
unique UID per namespace.

The `hello` app has been updated to use it. And this will have to be
generalized to all arnold apps.