🔧(redis) speed up redis startup time
The readinessProbe declared in the redis deployment has an
initialDelaySeconds of 60s. As a result, even if the pod is started
successfully, we have to wait 60s to be able to reach it through a
Service.

It slows down deployment of other apps that depends on redis, and CI
jobs are also impacted by this.