🐛(redirect) remove labels from redirect-nginx deployment
We have a fatal error if we try to re-deploy the `redirect` core app:
the `redirect-nginx` Deployment cannot to be updated.

This is due to the fact that the `.spec.selector.matchLabels` field is
Immutable in a Deployment and that we try to patch it with different
values.

Since the `redirect` core_apps is not a blue/green application and
that the `redirect-nginx` deployment has a static name, the dynamic
labels have been removed to fix this.