✨(core) define a default storage class with ansible variable
A default StorageClass is not necessarily defined on a k8s cluster.
When this is the case, the PersistentVolumeClaim created by arnold are
stuck in the "Pending" state.

In order to support more use cases, new ansible variables has been
introduced to define the name of the storage class to use by default,
depending on the accessMode requested. All applications have been
updated to use them.