🔧(apps) allow to switch the acme environment easily
We want to be able to switch the acme environment used to generate the
TLS certificate of arnold applications. It requires manual actions in
the current state of the project, due to caching issues and entities
automatically managed by cert-manager.

To make the switch process easier, we include the acme environment
(live or staging) in the name of the issuer and in the name of the TLS
secret names.