🔧(redis-sentinel) update redis-sentinel to be compatible with k8s
We want to deploy the redis-sentinel app on a kubernetes cluster.
Since all the work is done by the redis-operator, the application
could have been deployed on kubernetes as-is. But a few improvements
were made : the securityContext is now consistent with the rest of the
apps and the variables redis_sentinel_user_id, redis_sentinel_group_id
and redis_sentinel_fsgroup_id have been removed. Also, the storage
class to use has been added.