🐛(core) trim templates before checking if we should ignore them
We are unable to deploy the forum on non-trashable environment type.
This is due to the fact that the mongodb deployment template contains
an extra-line at the end of file, and is not ignored by the "Ignore
empty rendered templates" task. It results in a fatal error during
deployment. The ansible task has been fixed to trim files and avoid
this kind of error.