⚡️(scripts) add docker registry to k3d local cluster
When using arnold in a CI context for third-party apps deployable as
trays, we often have to build a Docker image of the current source code
state and test this code in a k8s context. To do so, we need a local
registry to push our newly built image that will be deployed.

To enable registry creation and attach it to the cluster, you should
define the K3D_ENABLE_REGISTRY=1 environment variable.