⚡️(elasticsearch) fix data volume permissions
While using RWO storage classes/providers (e.g. `cinder-csi`), PVC may
have wrong ownership and the user running in the container is not
allowed to write to the elasticsearch data volume (can belong to
`root`). Properly setting the `fsGroup` in the security context of the
pod fixes the issue.

To add more flexibility it is now also possible to also set a container
UID or GID for elasticsearch pods.

For reference, see k8s documentation:
https://kubernetes.io/docs/tasks/configure-pod-container/security-context/