🐛(ashley) allow adding alternative routes
Some browsers have started blocking third party cookies from within
an iframe. As a result Ashley is broken is broken when integrate in
an LMS page.

The solution is to host Ashley on a domain that is a sibling of the
domain on which the LMS itself is hosted. We can do that by allowing
to declare alternative routes for the ashley app.