🔧(elasticsearch|kibana) upgrade default releases to 7.16.2
With the recent log4j CVE, we've considered upgrading default
elasticsearch suite releases to the most recent patched version.

Note that the elasticsearch default image is now the official
distribution.

Note that, to avoid a deprecation warning, we had to make the
`server.publicBaseUrl` Kibana's setting configurable using the latest
release.