👷(circleci) migrate to machine image stable release
CircleCI will soon deprecate default base image for the machine
executor. We need to explicitly switch to the latest image stable
release.

For more information about this migration please, refer to the migration
guide:

https://circleci.com/docs/2.0/images/linux-vm/14.04-to-20.04-migration/

Note that installed curl release in this image now switch to HTTP/2 when
supported by the web server, so we needed to fix our CI script.