⬆️(project) upgrade base k3d and kubectl dependencies
The 5.x+ k3d releases break our k3d cluster initialization script, we
need to stay the most up-to-date we can.

Prosody seems to have permission issues on created PVCs. Using the "sql"
backend instead seems to fix the issue that is mostly related to a
k3d-circleci incompatibility for some applications.