♻️(marsha) create a CronJobList resource managing all CronJobs
The number of cronjobs is growing and they all follow the same pattern.
We define a name, a schedule, a command, resources and the template is
always the same. We now define an array of cornjobs in the configuration,
containing all listed parameters before and we loop again to declare
every CronJob in the CronJobList resource.