✨(marsha) manage service worker path in nginx
The service worker we are using must be served at the root of the
application. But we are managing it as a static file and it is saved
with the other static file managed by django. We need to add a rule
in the nginx conf to access it from the root.